+++ Data/Scripts/OreDetectorReforged/Config.cs
@@ -1,0 +1,12 @@
+﻿namespace OreDetectorReforged
+{
+    public class Config
+    {
+        public byte voxelContentMinPlanet = 96;
+        public byte voxelContentMinNotPlanet = 96;
+        public int biomeRepeatDivisor = 25;
+        public float maxRangeSmall = 1e9f;
+        public float maxRangeLarge = 1e9f;
+        public int maxCount = 100;
+    }
+}